Don't allow deletion of providers if projects active
Fixes #825, and a few express deprecation fixes and quotes.

Basically returns a 403 and a description stating which projects are using
the given provider.

This is probably an stop-gap solution, I would like to see the
projects listen by name in the accounts page and the delete button disabled
with a note stating that projects must first be deleted before the provider
can be deleted.