Crash on refreshing account on missing plugin
Steps to reproduce:
Have an instance of strider with say, a
gitlab account and a github account
configured. Now for some reason, go to a
state where the gitlab plugin is not installed.

(In my case, this was a problem with the npm
cache/vboxfs/or something else -- was resolved
with doing a clean re-clone of the strider repo
and npm link with a fresh copy of the strider-github
repo --- the relevant part here being, the strider-gitlab
plugin was *NOT* installed.)

So, the Mongo DB had an account configured for gitlab,
but the plugin was not present. In such a situation,
trying to click on the "Refresh Account" on the Github
page was resulting in a crash -
TypeError: Cannot read property 'listRepos' of undefined
        on strider/lib/routes/projects.js:78
Fix:
For the moment, we are using req.flash to show the
appropriate error message.

However, in case of multiple problems in the parallel
tasks, only one of the errors is being shown in the GUI.
This is scope for enhancement.